Move OpenTelemetrySdkBuildItem to core deployment module
