+++ core/deployment/src/main/java/io/quarkus/deployment/builditem/OpenTelemetrySdkBuildItem.java
@@ -1,4 +1,4 @@
-package io.quarkus.arc.deployment;
+package io.quarkus.deployment.builditem;
 
 import java.util.Optional;
 
@@ -6,9 +6,41 @@ import io.quarkus.builder.item.SimpleBuildItem;
 import io.quarkus.runtime.RuntimeValue;
 
 /**
- * @deprecated Use {@link io.quarkus.deployment.builditem.OpenTelemetrySdkBuildItem} instead.
+ * Indicates that the OpenTelemetry SDK is present and carries the build-time enabled state
+ * for each telemetry signal (tracing, metrics, logging) and a check if the OTel SDK is enabled at runtime.
+ * This can be used to decide if OTel instrumentation needs to be instantiated, or activated.
+ * <p>
+ * Produced by the OpenTelemetry extension's deployment module. Other extensions should consume
+ * it as an {@link Optional} so they still work when the OpenTelemetry extension is not on the
+ * classpath.
+ * <p>
+ * The build-time flags reflect the following configuration properties:
+ * <ul>
+ * <li>{@code quarkus.otel.traces.enabled} — tracing (defaults to {@code true} when the SDK is enabled)</li>
+ * <li>{@code quarkus.otel.metrics.enabled} — metrics (defaults to {@code false})</li>
+ * <li>{@code quarkus.otel.logs.enabled} — logging (defaults to {@code false})</li>
+ * </ul>
+ * Each signal is only active when both its own flag <em>and</em> the top-level
+ * {@code quarkus.otel.enabled} flag are {@code true}.
+ *
+ * <h2>Usage example</h2>
+ *
+ * <pre>{@code
+ * &#64;BuildStep
+ * void configureTracing(Optional<OpenTelemetrySdkBuildItem> otelSdk, ...) {
+ *     if (otelSdk.isPresent() && otelSdk.get().isTracingBuildTimeEnabled()) {
+ *         // register tracing integration
+ *     }
+ * }
+ * }</pre>
+ *
+ * The convenience method {@link #isOtelSdkEnabled(Optional)} extracts the runtime-enabled
+ * {@link RuntimeValue} when the build item is present:
+ *
+ * <pre>{@code
+ * Optional<RuntimeValue<Boolean>> runtimeEnabled = OpenTelemetrySdkBuildItem.isOtelSdkEnabled(otelSdk);
+ * }</pre>
  */
-@Deprecated(forRemoval = true)
 public final class OpenTelemetrySdkBuildItem extends SimpleBuildItem {
 
     private final boolean tracingBuildTimeEnabled;
